Design: 메인 페이지 디자인 수정
