+++ src/style/mainPage.scss
@@ -1,21 +1,60 @@
 @import './_reset';
 
 .swiper {
-  background-color: aquamarine;
-  height: 630px;
-  .swiper-slide{
+  background-color: #f5f5f5;
+  height: 660px;
+  .swiper-slide {
     img {
-    width: 800px;
-    height: 630px;
-    background-color: orange;
+      width: 44vw;
+      height: 600px;
+      background-color: #bcbfda;
     }
   }
 }
+.swiper-button-prev,
+.swiper-button-next {
+  &::after {
+    color: #2d2f43;
+    position: absolute;
+  }
+}
 
+.swiper-button-prev {
+  &::after {
+    left: 23vw;
+  }
+}
+.swiper-button-next {
+  &::after {
+    right: 23vw;
+  }
+}
+
+.swiper-pagination {
+  .swiper-pagination-bullet {
+    width: 10px;
+    height: 10px;
+    background-color: #999;
+    opacity: 1;
+    margin: 0 9px !important;
+  }
+  .swiper-pagination-bullet-active {
+    transform: scale(1.3);
+    border-radius: 5px;
+    background-color: transparent;
+    border: 1px solid #2d2f43;
+  }
+}
 .view-more {
-  width: 800px;
-  margin-left: 700px;
-  text-align: right;
+  color: #2d2f43;
+  font-size: 1.2rem;
+  position: absolute;
+  z-index: 2;
+  top: 750px;
+  right: 28vw;
+  .view-more-arrow {
+    margin-left: 10px;
+  }
 }
 
 // 페이지 상단으로 위치하는 버튼
